Created a gameboard and cell factory functions
The GameBoard factory creates a private two
dimensional array that represeants grid that plays
are made on. The inner arrays repreasent rows,
each row contains a series of cell. Vertically
contiguous cells makeup a column.

It returns a method that allows users update the
board state and another function that prits the
board state.

The Cell factory creates a private object for
recording a player's mark. It returns a method
that allows a player to add their mark to a cell
in the gameboard and another function that checks
the state of a cell in the gameboard.