Added a factory function to control the game flow
The Game() function creates an array with two
players, these are objects that have a
name and mark. The factory encapsulates
the following methods:

SwitchPlayer - Gives control to the next player

MakePlay - Adds the active player's mark to the
gameboard.

GetActivePlayer - Returns the player who currently
has control.

The active player is set to player one at the
offset of the game. The player is switched after
every valid play.