Changed the location of two functions
I moved Cell() outside of the Game
factory and encapsulated in the GameBoard module.
It is a more appropriate location because Cell()
is only invoked in GameBoard.

I moved PrintBoard form the GameBoard module to
the Game factory function because I want to allow
users to print board at any time throughout the
game.

FindWinner is now able to determine when there's
a winning row or column and return the winner.